api/requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
babel==2.17.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
Flask==3.1.0
flask-cors==5.0.0
Flask-Babel==4.0.0
Flask-Login==0.6.3
Flask-Sitemap==0.4.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.66.0
greenlet==3.2.4
grpcio==1.75.0rc1
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
Markdown==3.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.0
pytesseract==0.3.13
python-bidi==0.6.6
python-docx==1.1.2
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.3
rsa==4.9
setuptools==75.6.0
soupsieve==2.6
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
youtube-transcript-api==0.6.3